pushswap final?

+++ src/validate_input.c
@@ -6,20 +6,20 @@
 /*   By: maroy <maroy@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/07 13:56:14 by maroy             #+#    #+#             */
-/*   Updated: 2023/06/26 20:05:21 by maroy            ###   ########.fr       */
+/*   Updated: 2023/07/13 16:04:53 by maroy            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/push_swap.h"
 
 //checks if the numbers are valid, returns OK if true, KO if false
-static int32_t	is_valid_numbers(char **numbers, int32_t size);
+static int8_t	is_valid_numbers(char **numbers, int32_t size);
 //checks if the numbers are duplicates, returns OK if true, KO if false
-static int32_t	has_duplicate_values(char **argv);
+static int8_t	has_duplicate_values(char **argv);
 //checks if the string is a valid int32_tegers, returns OK if true, KO if false
-static int32_t	is_valid_int32_teger(char *str);
+static int8_t	is_valid_int32_teger(char *str);
 
-int32_t	validate_input(int32_t argc, char **argv)
+int8_t	validate_input(int32_t argc, char **argv)
 {
 	char	**split_num;
 	int32_t	exit_status;
@@ -43,7 +43,7 @@ int32_t	validate_input(int32_t argc, char **argv)
 	return (is_valid_numbers(argv + 1, argc - 1));
 }
 
-static int32_t	is_valid_numbers(char **numbers, int32_t size)
+static int8_t	is_valid_numbers(char **numbers, int32_t size)
 {
 	int32_t	i;
 
@@ -59,7 +59,7 @@ static int32_t	is_valid_numbers(char **numbers, int32_t size)
 	return (OK);
 }
 
-static int32_t	has_duplicate_values(char **argv)
+static int8_t	has_duplicate_values(char **argv)
 {
 	int32_t	i;
 	int32_t	j;
@@ -79,7 +79,7 @@ static int32_t	has_duplicate_values(char **argv)
 	return (OK);
 }
 
-static int32_t	is_valid_int32_teger(char *str)
+static int8_t	is_valid_int32_teger(char *str)
 {
 	int32_t	i;
 	long	num;
